# src/schemas/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Electric Disturbance Events",
    "description": "This API retrieves read only data related to electric disturbance events in the U.S.A.",
    "version": "1.0"
  },
  "externalDocs": {
    "description": "Find out more about this project",
    "url": "https://github.com/MyCodeForge/Electric-Emergency-and-Disturbance-Dashboard-Portfolio-Project"
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1"
    }
  ],
  "tags": [
    {
      "name": "disturbance-events",
      "description": "Data on disturbance events."
    }
  ],
  "paths": {
    "/disturbance-events/": {
      "get": {
        "tags": [
          "disturbance-events"
        ],
        "summary": "Get disturbance events.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisturbanceEvent"
                }
              }
            }
          }
        }
      }
    },
    "/disturbance-events/month/{month}": {
      "get": {
        "tags": [
          "disturbance-events"
        ],
        "summary": "Get disturbance events by month.",
        "description": "Multiple months can be provided with comma separated strings",
        "operationId": "getDisturbanceEventsByMonth",
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "description": "Status values that need to be considered for filter",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "none",
              "enum": [
                "January",
                "February",
                "March",
                "April",
                "May",
                "June",
                "July",
                "August",
                "September",
                "October",
                "November",
                "December"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisturbanceEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Disturnace events not found"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/disturbance-events/region/{region}": {
      "get": {
        "tags": [
          "disturbance-events"
        ],
        "summary": "Get disturbance events by regions.",
        "description": "Multiple regions can be provided with comma separated strings",
        "operationId": "getDisturbanceEventsByRegion",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "schema": {
              "type": "string",
              "default": "none",
              "enum": [
                "FRCC",
                "MRO",
                "N/A",
                "NPCC",
                "PR",
                "RF",
                "RFC",
                "SERC",
                "TRE",
                "WECC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisturbanceEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Disturnace events not found"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/disturbance-events/eventtype/{eventtype}": {
      "get": {
        "tags": [
          "disturbance-events"
        ],
        "summary": "Get disturbance events by event types.",
        "description": "Multiple event types can be provided with comma separated strings",
        "operationId": "getDisturbanceEventsByEventType",
        "parameters": [
          {
            "name": "eventtype",
            "in": "path",
            "description": "Status values that need to be considered for filter",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string",
              "default": "none",
              "enum": [
                "Cyber Attack",
                "Cyber Event",
                "Distribution Interruption",
                "Fuel Supply Deficiency",
                "Generation Inadequacy",
                "Islanding",
                "Natural Disaster",
                "Other",
                "Public Appeal",
                "Physical Attack",
                "Sabotage",
                "Severe Weather",
                "Suspicious Activity",
                "System Operations",
                "Transmission Interruption",
                "Vandalism",
                "Weather"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisturbanceEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Disturnace events not found"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/regions/summary/": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "Get a high-level overview of regional impact.",
        "description": "Returns a summary of events by NERC region (e.g., count of events, average demand loss in mw).",
        "operationId": "getDisturbanceEventsSummaryByRegion",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionSummary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DisturbanceEvent": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "examples": [
              "March"
            ]
          },
          "date_event_began": {
            "type": "string",
            "examples": [
              "3/11/2016"
            ]
          },
          "time_event_began": {
            "type": "string",
            "examples": [
              "4:00 AM"
            ]
          },
          "date_of_restoration": {
            "type": "string",
            "examples": [
              "3/13/2016"
            ]
          },
          "time_of_restoration": {
            "type": "string",
            "examples": [
              "11:59 AM"
            ]
          },
          "area_affected": {
            "type": "string",
            "examples": [
              "Washington: Kitsap County, King County, Whatcom County, Island County, Skagit County;"
            ]
          },
          "nerc_region": {
            "type": "string",
            "examples": [
              "WECC"
            ]
          },
          "alert_criteria": {
            "type": "string",
            "examples": [
              "Suspected Physical Attack"
            ]
          },
          "event_type": {
            "type": "string",
            "examples": [
              "Sabotage"
            ]
          },
          "demand_loss_in_mw": {
            "type": "string",
            "examples": [
              "Unknown"
            ]
          },
          "customers_affected": {
            "type": "string",
            "examples": [
              415103
            ]
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "SummaryMetrics": {
        "type": "object",
        "properties": {
          "average": {
            "type": "number",
            "format": "double",
            "example": 100.2,
            "description": "Average value"
          },
          "min": {
            "type": "number",
            "example": 50,
            "description": "Minimum value"
          },
          "max": {
            "type": "number",
            "example": 150,
            "description": "Maximum value"
          }
        }
      },
      "AffectedAreaSummary": {
        "type": "object",
        "properties": {
          "area": {
            "type": "string",
            "example": "California",
            "description": "Affected area name"
          },
          "event_count": {
            "type": "integer",
            "example": 10,
            "description": "Number of events in the area"
          }
        }
      },
      "RegionSummary": {
        "type": "object",
        "properties": {
          "from_date": {
            "type": "string",
            "format": "date",
            "example": "2024-03-15",
            "description": "Start date for the summary period"
          },
          "to_date": {
            "type": "string",
            "format": "date",
            "example": "2024-03-22",
            "description": "End date for the summary period"
          },
          "total_events": {
            "type": "integer",
            "example": 123,
            "description": "Total number of disturbance events"
          },
          "demand_loss_in_mw": {
            "$ref": "#/components/schemas/SummaryMetrics"
          },
          "customers_affected": {
            "$ref": "#/components/schemas/SummaryMetrics"
          },
          "event_types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Cyber Attack",
              "Severe Weather"
            ],
            "description": "List of event types"
          },
          "average_downtime": {
            "type": "number",
            "format": "double",
            "example": 24.5,
            "description": "Average downtime in hours"
          },
          "downtime_unit": {
            "type": "string",
            "example": "hours",
            "description": "Unit of downtime measurement"
          },
          "affected_areas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AffectedAreaSummary"
            },
            "description": "Summary of affected areas"
          },
          "no_events": {
            "type": "boolean",
            "example": false,
            "description": "Indicates if there were no events in the period"
          }
        }
      }
    }
  }
}